Make intermediate result call only when solution is valid
The CF intermediate result consumer was previously called whenever the new score was better than the previous one.
This could, however, happen even when the new proposed solution was _not_ a valid counterfactual.
This change makes the intermediate consumer to only be called if the proposed solution is a valid CF.