FAI-658: clear unnecessary settings
log4j2.xml was removed. The reason is that the config file  was upload into container but the log file does not contain more messages
aliases of containers is not supported in network by DMP. URL with alias does not work
repository of Docker was removed because it is set by CI workflow.