kie-issues#525
Closes: https://github.com/kiegroup/kie-issues/issues/525

The issue is related to the `FormDisplayer` component and its initialization.

Lets have a look firstly into `FormDsiplayer`. There we will find this code, for `html` forms.
```
<HtmlFormRenderer source={source} resources={resources} />
```

In the render logic, we can not execute ^ sooner than `source` is set, otherwise error will be thrown. Thus adding lgical check for `source` to be defined.