Add a null check on milestones attribute in ProcessDetails component
Building and running the data-index service and the management console based on master branch, it seems that milestones attribute can be now null instead of being an empty array as before. 

I have the feeling that it is following this big refactor on the serialization of the process instance feed between the kogito runtime and the data-index service:
https://github.com/apache/incubator-kie-kogito-apps/commit/ffc8500bbdfb979598392719d3b629a673ee5bdf

A simple optional chaining operator is resolving the issue.